+++ tests/testthat/test-newMultispeciesParams.R
@@ -48,7 +48,20 @@ test_that("Errors are reported", {
                  "min_w_pp must be larger than min_w")
 })
 
-## setParams ----
+# setParams ----
 test_that("setParams can leave params unchanged", {
     expect_unchanged(setParams(NS_params), NS_params)
+})
+
+test_that("setParams handles change in w_max", {
+    params <- NS_params
+    # Check that ft_mask is recalculated correctly
+    params@species_params$w_max[1] <- 1000
+    params <- setParams(params)
+    expect_equal(sum(params@ft_mask[1, ]), 205)
+    
+    # Check that error is thrown if w_max is too large
+    params@species_params$w_max[1] <- max(params@w) + 10
+    expect_error(setParams(params),
+                 "The maximum weight of a species is larger than")
 })